pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "digital-twin"
version = "0.1.0"
description = "Documentation intake service for diecast machine digital twin"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "fastapi>=0.115.0",
  "uvicorn[standard]>=0.30.0",
  "python-multipart>=0.0.9",
  "pypdf>=4.3.1",
  "pdfplumber>=0.11.4",
  "python-dotenv>=1.0.1",
  "google-genai>=1.0.0",
]

[tool.setuptools.packages.find]
where = ["src"]
